pyproject.toml
[project]
name = "sparkathon-agent"
version = "0.1.0"
description = "Walmart Multi-Agent Supply Chain Optimization System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.83.0",
    "crewai-tools>=0.13.0",
    "openai>=1.91.0",
    "python-dotenv>=1.1.1",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "asyncio-mqtt>=0.16.0",
    "geopy>=2.4.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
]
